+++ code/src/nuvla/job/distributions/notify_coupon_end.py
@@ -1,0 +1,37 @@
+# -*- coding: utf-8 -*-
+
+import json
+import logging
+from ..util import override
+from ..distributions import distribution
+from ..distribution import DistributionBase
+
+
+@distribution('notify_coupon_end')
+class NotifyCouponEndJobsDistribution(DistributionBase):
+    DISTRIBUTION_NAME = 'notify_coupon_end'
+
+    def __init__(self, distributor):
+        super(NotifyCouponEndJobsDistribution, self).__init__(
+            self.DISTRIBUTION_NAME,
+            distributor)
+        self.collect_interval = 3600  # 1 hour
+        self._start_distribution()
+
+    def list_coupons(self):
+        return self.distributor.api.hook('list-coupons').data
+
+    def coupon_ids(self):
+        try:
+            return [coupon.get('id') for coupon in self.list_coupons() if
+                    coupon.get('id')]
+        except Exception as ex:
+            logging.error(f'Failed to list coupons: {ex}')
+            return []
+
+    @override
+    def job_generator(self):
+        for coupon_id in self.coupon_ids():
+            yield {'action': self.DISTRIBUTION_NAME,
+                   'target-resource': {'href': 'hook/notify-coupon-end'},
+                   'payload': json.dumps({'coupon': coupon_id})}